Exclusion filters and rule prefilter
