+++ CMakeLists.txt
@@ -71,6 +71,7 @@ set(LIBDDWAF_SOURCE
     ${libddwaf_SOURCE_DIR}/src/object.cpp
     ${libddwaf_SOURCE_DIR}/src/manifest.cpp
     ${libddwaf_SOURCE_DIR}/src/object_store.cpp
+    ${libddwaf_SOURCE_DIR}/src/condition.cpp
     ${libddwaf_SOURCE_DIR}/src/rule.cpp
     ${libddwaf_SOURCE_DIR}/src/rule_data_dispatcher.cpp
     ${libddwaf_SOURCE_DIR}/src/ruleset_info.cpp
