+++ src/context.hpp
@@ -15,6 +15,7 @@
 #include <config.hpp>
 #include <utils.h>
 #include <obfuscator.hpp>
+#include <rule_prefilter.hpp>
 #include <ruleset.hpp>
 
 namespace ddwaf
@@ -23,9 +24,10 @@ namespace ddwaf
 class context
 {
 public:
-    context(const ddwaf::ruleset &ruleset, const ddwaf::config &config):
+    context(ddwaf::ruleset &ruleset, const ddwaf::config &config):
         ruleset_(ruleset), config_(config),
-        store_(ruleset_.manifest, config_.free_fn)
+        store_(ruleset_.manifest, config_.free_fn),
+        prefilter(ruleset)
     {
         status_cache_.reserve(ruleset_.rules.size());
     }
@@ -39,16 +41,20 @@ public:
     DDWAF_RET_CODE run(const ddwaf_object&, optional_ref<ddwaf_result> res, uint64_t);
 
 protected:
-    bool run_collection(const std::string& name,
-        const ddwaf::rule_ref_vector& collection,
+    bool run_rules(const ddwaf::rule_ref_vector& rules,
         event_serializer& serializer,
         ddwaf::timer& deadline);
 
     bool is_first_run() const { return status_cache_.empty(); }
 
-    const ddwaf::ruleset &ruleset_;
+    ddwaf::ruleset &ruleset_;
     const ddwaf::config &config_;
     ddwaf::object_store store_;
+    ddwaf::rule_prefilter prefilter;
+
+    // Cache collections to avoid processing once a result has been obtained
+    // TODO: strings should be replaced by a scalar.
+    std::unordered_set<std::string> collection_cache_;
 
     // If we have seen a match, the value will be true, if the value is present
     // and false it means we executed the rule and it did not match.
